create checkout_mixin.rb for handling stateless payment amount calculations
rename checkout -> checkout_mixin and remove unnecessary rails autoload change.
