+++ config/application.rb
@@ -65,6 +65,5 @@ module Crowdhoster
 
     config.processing_fee_percentage = 2.9
     config.processing_fee_flat_cents = 30
-
   end
 end
